refactor: refactoring dtos
