+++ backend/factory-management/src/main/java/com/projetada/factorymanagement/models/Product.java
@@ -1,6 +1,5 @@
 package com.projetada.factorymanagement.models;
 
-import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 
 import java.io.Serializable;
